requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiosignal==1.3.2
annotated-types==0.7.0
antiorm==1.2.1
anyio==4.9.0
asttokens==3.0.0
attrs==25.3.0
blinker==1.9.0
certifi==2025.1.31
click==8.1.8
comm==0.2.2
db==0.1.1
db-sqlite3==0.0.1
debugpy==1.8.13
decorator==5.2.1
deprecation==2.1.0
executing==2.2.0
Flask==3.1.0
frozenlist==1.5.0
gotrue==2.11.4
greenlet==3.1.1
h11==0.14.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.7
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
ipykernel==6.29.5
ipython==9.0.2
ipython_pygments_lexers==1.1.1
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jupyter_client==8.6.3
jupyter_core==5.7.2
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
multidict==6.2.0
nest-asyncio==1.6.0
packaging==24.2
parso==0.8.4
pexpect==4.9.0
platformdirs==4.3.7
postgrest==0.19.3
prompt_toolkit==3.0.50
propcache==0.3.0
psutil==7.0.0
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pyzmq==26.3.0
realtime==2.4.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.39
stack-data==0.6.3
storage3==0.11.3
StrEnum==0.4.15
supabase==2.14.0
supafunc==0.9.3
tornado==6.4.2
traitlets==5.14.3
typing_extensions==4.12.2
wcwidth==0.2.13
websockets==14.2
Werkzeug==3.1.3
yarl==1.18.3
gunicorn==20.1.0
